servers/onepassword-mcp/pyproject.toml
[project]
name = "onepassword-mcp"
version = "0.1.0"
description = "MCP server for 1Password using official SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "onepassword-sdk>=0.1.0",
]

[project.scripts]
onepassword-mcp = "onepassword_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/onepassword_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
